+++ apps/demo-app/src/App.vue
@@ -8,12 +8,11 @@ import {
   OnyxMenuItem,
   OnyxNavBar,
   OnyxNavButton,
+  OnyxNavItem,
   OnyxToast,
   OnyxUserMenu,
   useThemeTransition,
-  type OnyxNavButtonProps,
 } from "sit-onyx";
-import { useTemplateRef, watch } from "vue";
 import { RouterView, useRoute, useRouter } from "vue-router";
 import onyxLogo from "./assets/onyx-logo.svg";
 import { useGridStore } from "./stores/grid-store";
@@ -22,25 +21,8 @@ const router = useRouter();
 const route = useRoute();
 const gridStore = useGridStore();
 
-const navItems = [
-  { label: "Home", link: "/" },
-  { label: "Form Demo", link: "/form-demo" },
-  { label: "Layout Demo", link: "/layout-demo" },
-  { label: "Grid Demo", link: "/grid" },
-  { label: "Data-Grid Demo", link: "/data-grid" },
-] satisfies OnyxNavButtonProps[];
-
 const { store: colorScheme } = useColorMode({ disableTransition: false });
 useThemeTransition(colorScheme);
-
-const navBar = useTemplateRef("navBarRef");
-
-watch(
-  () => route.path,
-  () => {
-    navBar.value?.closeMobileMenus();
-  },
-);
 </script>
 
 <template>
@@ -55,25 +37,26 @@ watch(
         /* the layout demo showcases all possible overlay features that AppLayout offers,
         including different nav bar behaviors which OnyxNavBar das not support yet
         so it currently has its own demo nav bar placed. */
-        route.path !== '/layout-demo'
+        route.path !== '/demos/layout'
       "
       #navBar
     >
       <OnyxNavBar
-        ref="navBarRef"
         app-name="Demo App"
         :logo-url="onyxLogo"
         show-back-button
         @back-button-click="router.back"
-        @app-area-click="router.push('/')"
       >
-        <OnyxNavButton
-          v-for="item in navItems"
-          :key="item.link"
-          v-bind="item"
-          :active="item.link === router.currentRoute.value.path"
-          @navigate="router.push"
-        />
+        <OnyxNavButton label="Home" link="/" />
+
+        <OnyxNavButton label="Demos">
+          <template #children>
+            <OnyxNavItem label="Form" link="/demos/form" />
+            <OnyxNavItem label="Layout" link="/demos/layout" />
+            <OnyxNavItem label="Grid" link="/demos/grid" />
+            <OnyxNavItem label="Data grid" link="/demos/data-grid" />
+          </template>
+        </OnyxNavButton>
 
         <template #contextArea>
           <OnyxUserMenu full-name="John Doe">