+++ apps/demo-app/src/views/HomeView.vue
@@ -224,7 +224,9 @@ const selectedDate = ref<DateValue>();
 
         <template v-if="show('OnyxLink')">
           <OnyxLink href="https://onyx.schwarz">External link</OnyxLink>
-          <OnyxLink href="/form-demo">Internal link (should be navigated with Vue Router)</OnyxLink>
+          <OnyxLink href="/demos/form">
+            Internal link (should be navigated with Vue Router)
+          </OnyxLink>
         </template>
 
         <SelectDemo