+++ packages/nuxt/playground/nuxt.config.ts
@@ -18,7 +18,7 @@ export default defineNuxtConfig({
     ],
   },
   onyx: {
-    theme: "kaufland",
+    theme: "lidl",
     i18n: {
       registerLocales: {
         "en-US": "en-US",