+++ packages/sit-onyx/src/components/OnyxNavBar/modules/OnyxMenuItem/types.ts
@@ -3,10 +3,11 @@ import type { WithLinkProp } from "../../../OnyxRouterLink/types";
 
 export type OnyxMenuItemProps = WithLinkProp & {
   /**
-   * Whether the menu item is currently active.
-   * This is the case if the given link is currently opened.
+   * Whether the item is currently active.
+   * This is the case if the page of the given link is currently open.
+   * If "auto" and a [router](https://onyx.schwarz/development/router.html) is provided, the active state will be determined automatically based on the current route.
    */
-  active?: boolean;
+  active?: boolean | "auto";
   /**
    * Whether the menu item is disabled.
    */