+++ packages/sit-onyx/src/components/OnyxNavBar/modules/OnyxNavButton/OnyxNavButton.ct.tsx
@@ -9,6 +9,7 @@ import OnyxBadge from "../../../OnyxBadge/OnyxBadge.vue";
 import OnyxNavItem from "../OnyxNavItem/OnyxNavItem.vue";
 import MobileComponentTestWrapper from "./MobileComponentTestWrapper.ct.vue";
 import OnyxNavButton from "./OnyxNavButton.vue";
+import RouterTestWrapperCt from "./RouterTestWrapper.ct.vue";
 
 const context = {
   /**
@@ -137,4 +138,82 @@ test.describe("Screenshot tests (mobile children)", () => {
       },
     },
   });
+});
+
+test("should behave correctly without link", async ({ mount }) => {
+  // ARRANGE
+  let clickEventCount = 0;
+  const component = await mount(<OnyxNavButton label="Label" onClick={() => clickEventCount++} />);
+
+  // ACT
+  await component.getByRole("menuitem", { name: "Label" }).click();
+
+  // ASSERT
+  expect(clickEventCount).toBe(1);
+});
+
+test("should behave correctly with link", async ({ mount, page }) => {
+  // ARRANGE
+  const component = await mount(<OnyxNavButton label="Label" link="#link" />);
+
+  // ACT
+  await component.getByRole("menuitem", { name: "Label" }).click();
+
+  // ASSERT
+  await expect(page).toHaveURL(/^http:\/\/localhost:\d*\/#link$/);
+});
+
+test("should behave correctly with nested children", async ({ mount, page }) => {
+  // ARRANGE
+  const component = await mount(
+    <OnyxNavButton label="Label">
+      <OnyxNavItem label="Nested item 1" link="#nested-1" />
+    </OnyxNavButton>,
+  );
+
+  // ACT
+  await component.getByRole("menuitem", { name: "Label" }).hover();
+  await component.getByRole("menuitem", { name: "Nested item 1" }).click();
+
+  // ASSERT
+  await expect(page).toHaveURL(/^http:\/\/localhost:\d*\/#nested-1$/);
+});
+
+test("should auto detect active state based on provided router", async ({ mount }) => {
+  // ARRANGE
+  const activeClassName = /onyx-nav-button--active/;
+
+  const component = await mount(RouterTestWrapperCt, {
+    props: {
+      currentRoute: "/test",
+      label: "Test label",
+      link: "/test",
+    },
+  });
+
+  // ASSERT
+  await expect(component, "should be active if current route matches").toHaveClass(activeClassName);
+
+  // ACT
+  await component.update({ props: { active: false } });
+
+  // ASSERT
+  await expect(
+    component,
+    "should not be active if current route matches but explicitly set false",
+  ).not.toHaveClass(activeClassName);
+
+  // ACT
+  await component.update({ props: { currentRoute: "/not-test", active: "auto" } });
+
+  // ASSERT
+  await expect(component, "should NOT be active if current route does not match").not.toHaveClass(
+    activeClassName,
+  );
+
+  // ACT
+  await component.update({ props: { active: true } });
+
+  // ASSERT
+  await expect(component, "should be active if explicitly set true").toHaveClass(activeClassName);
 });