+++ packages/sit-onyx/src/composables/useLenientMaxLengthValidation.spec.ts
@@ -8,7 +8,7 @@ import {
 } from "./useLenientMaxLengthValidation";
 
 vi.mock("vue", async (importOriginal) => {
-  const module: typeof vue = await importOriginal();
+  const module = await importOriginal<typeof import("vue")>();
 
   return {
     ...module,