+++ packages/sit-onyx/src/composables/useLink.ts
@@ -1,11 +1,22 @@
-import { inject, type InjectionKey } from "vue";
+import { computed, inject, unref, type InjectionKey, type Ref } from "vue";
+import type { SharedLinkProps } from "../components/OnyxRouterLink/types";
 import { isExternalLink } from "../utils";
+import { extractLinkProps } from "../utils/router";
 
 /**
- * Internal behavior of the `OnyxRouterLink`. Should be used for navigation. Considers the user provided router.
+ * Internal behavior of the `OnyxRouterLink`. Should be used for navigation. Considers the user provided [router](https://onyx.schwarz/development/router.html).
  */
 export const useLink = () => {
   const router = inject(ROUTER_INJECTION_KEY, undefined);
+
+  /**
+   * Currently active route (if router is provided).
+   */
+  const currentRoute = computed(() => {
+    if (!router) return;
+    const route = unref(router.currentRoute);
+    return typeof route === "string" ? { path: route } : route;
+  });
 
   /**
    * Handles the navigation with the user provided router if available and the link is internal.
@@ -37,7 +48,22 @@ export const useLink = () => {
     }
   };
 
-  return { navigate };
+  /**
+   * Checks whether the given path is currently active based on the currently active route.
+   */
+  const isActive = computed(() => {
+    return (link?: string | SharedLinkProps) => {
+      if (!currentRoute.value || link == undefined) return false;
+
+      const href = normalizeHref(extractLinkProps(link).href);
+      const path = normalizeHref(currentRoute.value.path);
+
+      if (href === "/") return path === href;
+      return path.startsWith(href);
+    };
+  });
+
+  return { navigate, currentRoute, isActive };
 };
 
 /**
@@ -62,12 +88,30 @@ const shouldHandleNavigation = (e: MouseEvent) => {
   return true;
 };
 
+/**
+ * Normalized the given href by:
+ * - convert "" to "/"
+ * - remove trailing slashes (only if href is not "/" itself)
+ * - trim whitespaces
+ */
+const normalizeHref = (href: string) => {
+  const value = href.trim();
+  if (value === "") return "/";
+  if (value === "/") return value;
+  return value.replace(/\/+$/, "");
+};
+
 export type ProvideRouterOptions = {
   /**
    * Programmatically navigate to a new URL.
    * @see https://router.vuejs.org/api/interfaces/Router.html#push-
    */
   push: (to: string) => void;
+  /**
+   * Currently active route.
+   * @see https://router.vuejs.org/api/interfaces/Router.html#currentRoute
+   */
+  currentRoute: Ref<string | { path: string }>;
 };
 
 export const ROUTER_INJECTION_KEY = Symbol() as InjectionKey<ProvideRouterOptions>;