+++ packages/sit-onyx/src/utils/plugin.ts
@@ -12,7 +12,8 @@ export type OnyxPluginOptions = {
    */
   i18n?: ProvideI18nOptions;
   /**
-   * Integration for [Vue Router](https://router.vuejs.org/)
+   * Integration for [Vue Router](https://router.vuejs.org/).
+   * @see https://onyx.schwarz/development/router.html
    */
   router?: ProvideRouterOptions;
 };